feat - added new extractor functionality and updated dependencies +semver: major
