+++ BetsBrasileiras/BetsBrasileiras.csproj
@@ -13,11 +13,11 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="CrispyWaffle" Version="9.1.165" />
-    <PackageReference Include="CrispyWaffle.I18n.PtBr" Version="9.1.165" />
+    <PackageReference Include="CrispyWaffle" Version="9.1.268" />
+    <PackageReference Include="CrispyWaffle.I18n.PtBr" Version="9.1.268" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
     <PackageReference Include="PdfPig" Version="0.1.9" />
-    <PackageReference Include="System.Text.Encoding.CodePages" Version="9.0.0" />
+    <PackageReference Include="System.Text.Encoding.CodePages" Version="9.0.1" />
   </ItemGroup>
 
   <ItemGroup>
