+++ BetsBrasileiras/Helpers/Constants.cs
@@ -1,5 +1,6 @@
 ﻿namespace BetsBrasileiras.Helpers;
 
+#pragma warning disable S1075
 internal static class Constants
 {
     /// <summary>
@@ -12,11 +13,12 @@ internal static class Constants
     /// The base URL.
     /// </summary>
     public const string BaseUrl =
-        "https://raw.githubusercontent.com/guibranco/BetsBrasileiras/main/data/bancos.json";
+        "https://raw.githubusercontent.com/guibranco/BetsBrasileiras/main/data/bets.json";
 
     /// <summary>
     /// The Secretaria Premios & Apostas URL
     /// </summary>
     public const string SpaUrl =
-        "https://www.gov.br/fazenda/pt-br/composicao/orgaos/secretaria-de-premios-e-apostas/lista-de-empresas/planilha-de-autorizadas-08-01-25.pdf";
+        "https://www.gov.br/fazenda/pt-br/composicao/orgaos/secretaria-de-premios-e-apostas/lista-de-empresas/lista-bets-14-01.pdf";
 }
+#pragma warning restore S1075
