+++ BetsBrasileiras/Helpers/Patterns.cs
@@ -2,9 +2,30 @@
 
 namespace BetsBrasileiras.Helpers;
 
+/// <summary>
+/// Class Patterns.
+/// </summary>
 internal static class Patterns
 {
-    public static readonly Regex SpaPattern = new Regex(
-        @"(?<applicationNumber>\d{4})\s+(?<applicationYear>\d{4})\s+(?<name>.+)\s+(?<document>\d{3}\.\d{3}\.\d{3}-\d{2})\s+(?<brand>.+)\s+(?<domain>.+)"
+    /// <summary>
+    /// The flags
+    /// </summary>
+    private static RegexOptions _flags =
+        RegexOptions.Compiled | RegexOptions.CultureInvariant | RegexOptions.IgnoreCase;
+
+    /// <summary>
+    /// The spa full line pattern
+    /// </summary>
+    public static readonly Regex SpaFullLinePattern = new(
+        @"(?<applicationNumber>\d{4})\/(?<applicationYear>\d{4})\s+(?<name>.+)\s+(?<document>\d{2}\.\d{3}\.\d{3}\/\d{4}-\d{2})\s+(?<brand>.+?)\s(?<domain>.+)$",
+        _flags
+    );
+
+    /// <summary>
+    /// The spa brand domain pattern
+    /// </summary>
+    public static readonly Regex SpaBrandDomainPattern = new(
+        @"(?<brand>.+?)\s(?<domain>.+)$",
+        _flags
     );
 }
