+++ BetsBrasileiras/Helpers/Reader.cs
@@ -1,13 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Globalization;
-using System.Linq;
 using System.Net.Http;
-using System.Text;
-using System.Threading.Tasks;
 using BetsBrasileiras.Dto;
 using CrispyWaffle.Serialization;
-using Microsoft.VisualBasic;
 using UglyToad.PdfPig;
 using UglyToad.PdfPig.DocumentLayoutAnalysis.TextExtractor;
 
@@ -16,6 +11,12 @@ namespace BetsBrasileiras.Helpers;
 internal class Reader
 {
     /// <summary>
+    /// The user agent.
+    /// </summary>
+    private const string UserAgent =
+        "BetsBrasileiras/1.0 (+https://github.com/guibranco/betsbrasileiras)";
+
+    /// <summary>
     /// Downloads the string.
     /// </summary>
     /// <param name="url">The URL.</param>
@@ -23,10 +24,7 @@ internal class Reader
     private static string DownloadString(string url)
     {
         using var client = new HttpClient();
-        client.DefaultRequestHeaders.Add(
-            "User-Agent",
-            "BetsBrasileiras/1.0 (+https://github.com/guibranco/bancosbrasileiros)"
-        );
+        client.DefaultRequestHeaders.Add("User-Agent", UserAgent);
         using var response = client.GetAsync(url).Result;
         using var content = response.Content;
         return content.ReadAsStringAsync().Result;
@@ -40,10 +38,7 @@ internal class Reader
     private static byte[] DownloadBytes(string url)
     {
         using var client = new HttpClient();
-        client.DefaultRequestHeaders.Add(
-            "User-Agent",
-            "BetsBrasileiras/1.0 (+https://github.com/guibranco/bancosbrasileiros)"
-        );
+        client.DefaultRequestHeaders.Add("User-Agent", UserAgent);
         return client.GetByteArrayAsync(url).Result;
     }
 
@@ -51,7 +46,6 @@ internal class Reader
     /// Downloads and parse PDF.
     /// </summary>
     /// <param name="url">The URL.</param>
-    /// <param name="system">The system.</param>
     /// <param name="callback">The callback.</param>
     /// <returns>List&lt;Bet&gt;.</returns>
     private static List<Bet> DownloadAndParsePdf(
@@ -64,7 +58,7 @@ internal class Reader
 
         try
         {
-            Logger.Log($"Downloading", ConsoleColor.Green);
+            Logger.Log($"Downloading SPA", ConsoleColor.Green);
             var data = DownloadBytes(url);
             reader = PdfDocument.Open(data);
         }
@@ -93,7 +87,7 @@ internal class Reader
     /// Loads the base.
     /// </summary>
     /// <returns>List&lt;Bet&gt;.</returns>
-    public List<Bet> LoadBase()
+    public static List<Bet> LoadBase()
     {
         Logger.Log("Downloading base", ConsoleColor.Green);
         var data = DownloadString(Constants.BaseUrl);
@@ -112,31 +106,39 @@ internal class Reader
         var result = new List<Bet>();
         var lines = page.Split("\n");
 
-        var spliced = new StringBuilder();
-
+        var previousLine = string.Empty;
         foreach (var line in lines)
         {
-            if (!Patterns.SpaPattern.IsMatch(line))
+            var currentLine = line.Trim('\r');
+            if (
+                !Patterns.SpaFullLinePattern.IsMatch(currentLine)
+                && !Patterns.SpaBrandDomainPattern.IsMatch(currentLine)
+            )
             {
-                spliced.Append($" {line}");
                 continue;
             }
 
-            Bet bet;
+            Bet bet = null;
 
-            if (!string.IsNullOrWhiteSpace(spliced.ToString()))
+            if (
+                !string.IsNullOrWhiteSpace(previousLine)
+                && Patterns.SpaBrandDomainPattern.IsMatch(currentLine)
+            )
             {
-                bet = ParseLineSpa(spliced.ToString().Trim());
-
-                if (bet != null)
-                {
-                    result.Add(bet);
-                }
-
-                spliced.Clear();
+                // TODO: Implement this when able to extract the content from the PDF
+                //bet = ParseLineSpa($"{previousLine} {currentLine}");
             }
-
-            bet = ParseLineSpa(line);
+            else if (Patterns.SpaFullLinePattern.IsMatch(currentLine))
+            {
+                bet = ParseLineSpa(currentLine);
+                //if (bet != null)
+                //{
+                //    previousLine = currentLine
+                //        .Replace(bet.Brand, "")
+                //        .Replace(bet.Domain, "")
+                //        .Trim();
+                //}
+            }
 
             if (bet != null)
             {
@@ -149,18 +151,18 @@ internal class Reader
 
     private Bet ParseLineSpa(string line)
     {
-        if (!Patterns.SpaPattern.IsMatch(line))
+        if (!Patterns.SpaFullLinePattern.IsMatch(line))
         {
             return null;
         }
 
-        var match = Patterns.SpaPattern.Match(line);
+        var match = Patterns.SpaFullLinePattern.Match(line);
 
         return new Bet
         {
             ApplicationNumberString = match.Groups["applicationNumber"].Value.Trim(),
             ApplicationYearString = match.Groups["applicationYear"].Value.Trim(),
-            Name = match.Groups["name"].Value.Trim(),
+            FiscalName = match.Groups["name"].Value.Trim(),
             Document = match.Groups["document"].Value.Trim(),
             Brand = match.Groups["brand"].Value.Trim(),
             Domain = match.Groups["domain"].Value.Trim(),
